+++ LBHTenancyAPI/UseCases/Search/Models/SearchTenancyResponse.cs
@@ -1,9 +1,10 @@
+using LBH.Data.Domain;
 using System.Collections.Generic;
 
 namespace LBHTenancyAPI.UseCases.Contacts.Models
 {
     public class SearchTenancyResponse
     {
-        public IList<TenancySearch> SearchResults { get; set; }
+        public IList<TenancyListItem> SearchResults { get; set; }
     }
 }
